Fix sheet request filtering and display
This commit fixes the sheet request filtering and display logic to
correctly show sent and received requests.

It updates the following:

- The `filtered_by` scope in `SheetRequest` model. - The partial
`_sheet_request.html.erb` to display sender/receiver. - The
`index.html.erb` to display the correct filters. - The redirect on the
controller to use flash messages - The link to sheet_requests