Update tests to use account_url and fix UI issues
- Update email verification and update password tests to use account_url
instead of root_url. - Remove the handle field from the profile update
test. - Fix the sign in test to assert sign in works. - Fix session
tests to confirm the session is destroyed - Fix sheets tests to sign in
and use the correct buttons. - Fix workout tests to sign in and use the
correct buttons. - Fix workout tests to pass the sheet in the url